biletwep/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
google-generativeai
scikit-learn==1.3.0
pandas==2.1.0
